update amplifier schematic and change overflow counter from uint8 to uint16

+++ code/crankshaft_angle/crankshaft_angle.ino
@@ -6,7 +6,7 @@
 #define TDC_TOOTH 1
 
 uint8_t tooth = 1;
-uint8_t tmrOvf = 0;
+uint16_t tmrOvf = 0;
 uint16_t prevPeriod = 0;
 
 uint16_t m_micros(void) {
